Add files via upload
In this file, I analyze data from the IMDB website, which contains a wealth of information about numerous films. All the films in the sample are science fiction (and may be combined with other genres). I begin by importing the data and performing various manipulations, and then create various descriptive statistics and graphs.